server/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==42.0.5
dataclasses-json==0.6.7
distro==1.9.0
ecdsa==0.19.0
elastic-transport==8.15.1
elasticsearch==8.15.1
exceptiongroup==1.2.0
fastapi==0.110.0
fastjsonschema==2.19.1
frozenlist==1.5.0
git-filter-repo==2.38.0
google-auth==2.36.0
googleapis-common-protos==1.66.0
# grpcio==1.68.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
httpx-sse==0.4.0
idna==3.6
iniconfig==2.0.0
jiter==0.8.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_core==5.7.2
jupytext==1.16.1
langchain==0.3.9
langchain-community==0.3.8
langchain-core==0.3.21
langchain-experimental==0.3.3
langchain-openai==0.2.10
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.2
langsmith==0.1.147
lz4==4.3.3
markdown-it-py==3.0.0
marshmallow==3.23.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
nbformat==5.10.3
numpy==1.26.4
openai==1.55.3
orjson==3.10.12
packaging==24.0
pandas==2.2.1
# pip install "pinecone-client[grpc]"
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
platformdirs==4.2.0
pluggy==1.5.0
propcache==0.2.0
proto-plus==1.25.0
# protobuf==4.25.5
protoc-gen-openapiv2==0.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyMySQL==1.1.0
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.17
pytz==2024.1
PyYAML==6.0.1
redis==5.2.0
referencing==0.34.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.18.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.29
starlette==0.36.3
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
toml==0.10.2
tomli==2.2.1
tqdm==4.66.4
traitlets==5.14.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.3
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
yarl==1.18.0
google-api-core==2.23.0
google-cloud-vision==3.8.1
grpcio==1.68.1
grpcio-status==1.68.1
protobuf==5.29.1
pycryptodome==3.21.0